# Simplify and fix watchHDAddrs concurrency.