Whilst not commented, the use of runtime.NumCPU and channels in the old code indicated an intention to speed up execution by utilizing multiple threads. This did not work in practice.

Utilizing errgroup instead of this hand-crafted concurrency code gives a less verbose result which scales over multiple threads properly.

Using a test wallet which registers for txn notifications 2.6 million addresses on startup:

Before this commit:
- `GOMAXPROCS=1    99.1 seconds`
- `GOMAXPROCS=2    97.3 seconds`
- `GOMAXPROCS=6    96.1 seconds`

After:
- `GOMAXPROCS=1    98.0 seconds`
- `GOMAXPROCS=2    49.5 seconds`
- `GOMAXPROCS=6    17.0 seconds`